refactor: move my home page to screens directory
